fix: removing some utils
